llm-agents-rag/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.9
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langgraph>=0.0.10
pypdf>=4.0.0
chromadb>=0.4.22
rank_bm25>=0.2.2
reportlab>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
